requirements.txt
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
requests==2.25.1
django-celery-results==2.0.1